get skewer script host from injected context
In PhantomJS we can't get /skewer's host from the script src (doesn't
have html tags) and we can't default to the context host because it
uses the file:// scheme, so let the injected context specify the host
with a global variable.

Other options are having skewer-mode send a header with the right host
or dynamically adding the host to skewer.js before injecting.

See: skeeto/skewer-mode#51