Fix skewer.log() when REPL is empty
There's a small skewer-repl bug.  I often clear the repl buffer via a
small interactive function (bound to a key) that calls comint-clear-
buffer.  However, there's a small bug in skewer-post-log which causes
messages via skewer.log() to not appear in the repl.

Here's what happens.  After the comint-clear-buffer clears the repl,
the prompt starts in the very first line.  When skewer.log() tries to
log to the repl it calls skewer-post-log which tries to move backwards
one line to insert before the prompt.  However, this fails because
there's no line before the current line and (backward-char) throws an
error (which appears to sometimes disconnects the connection to the
browser) and the text is not logged to the repl.